added input validation in the backend

+++ Application/Backend/routes/user.js
@@ -66,6 +66,7 @@ router.get("/", function (req, res, next) {
 router.post("/validate", function (req, res) {
   const { email, password } = req.body;
 
+  // to check that all user information is valid
   const isValidEmail = email !== undefined && typeof email == "string";
   const isValidPassword = password !== undefined && typeof password == "string";
 
@@ -86,33 +87,52 @@ router.post("/validate", function (req, res) {
         }
       }
     });
+  } else {
+    res.status(400).send("Invalid Input!");
   };
 });
 
 router.get('/watchlist', function (req, res, next) {
 
-  const userID = req.query.userID
-  const queryString = `SELECT * FROM movies where Movie_ID in (select Movie_ID from Watchlist where userID = ${userID})`;
-  connection.query(queryString, (error, results) => {
-    if (error) {
-      res.send(error);
-    } else {
-      res.send(results);
-    }
-  });
+  const { userID } = req.query;
+
+  // to check that all watchlist information is valid
+  const isValidUserID = userID !== undefined && typeof userID == "string";
+
+  if (isValidUserID) {
+    const queryString = `SELECT * FROM movies where Movie_ID in (select Movie_ID from Watchlist where userID = ${userID})`;
+    connection.query(queryString, (error, results) => {
+      if (error) {
+        res.send(error);
+      } else {
+        res.send(results);
+      }
+    });
+  } else {
+    res.status(400).send("Invalid Input!");
+  };
 });
 
 router.get('/watchlist/check', function (req, res, next) {
   const userID = req.query.userID
   const mid = req.query.mid
-  const queryString = `SELECT * FROM watchlist WHERE userID = ${userID} and Movie_ID = ${mid}`;
-  connection.query(queryString, (error, results) => {
-    if (error) {
-      res.send(error);
-    } else {
-      res.send(results);
-    }
-  });
+
+  // to check that all watchlist information is valid
+  const isValidUserID = userID !== undefined && typeof userID == "string";
+  const isValidMid = mid !== undefined && typeof mid == "string";
+
+  if (isValidUserID && isValidMid) {
+    const queryString = `SELECT * FROM watchlist WHERE userID = ${userID} and Movie_ID = ${mid}`;
+    connection.query(queryString, (error, results) => {
+      if (error) {
+        res.send(error);
+      } else {
+        res.send(results);
+      }
+    });
+  } else {
+    res.status(400).send("Invalid Input!");
+  };
 });
 
 router.post('/watchlist/add', function (req, res, next) {
@@ -121,17 +141,23 @@ router.post('/watchlist/add', function (req, res, next) {
     mid,
   } = req.body;
 
-  // add validators to check that everything is a string
-  
-  const queryString = `INSERT into watchlist (userID,Movie_ID) VALUES (${userID}, ${mid})`;
+  // to check that all watchlist information is valid
+  const isValidUserID = userID !== undefined && typeof userID == "string";
+  const isValidMid = mid !== undefined && typeof mid == "string";
 
-  connection.query(queryString, (error, results) => {
-    if (error) {
-      res.status(400).send('Database could not insert into watchlist');
-    } else {
-      res.send("Added new watchlist entry");
-    }
-  });
+  if (isValidUserID && isValidMid) {
+    const queryString = `INSERT into watchlist (userID,Movie_ID) VALUES (${userID}, ${mid})`;
+
+    connection.query(queryString, (error, results) => {
+      if (error) {
+        res.status(400).send('Database could not insert into watchlist');
+      } else {
+        res.send("Added new watchlist entry");
+      }
+    });
+  } else {
+    res.status(400).send("Invalid Input!");
+  };
 });
 
 router.post('/watchlist/remove', function (req, res, next) {
@@ -139,17 +165,26 @@ router.post('/watchlist/remove', function (req, res, next) {
     userID,
     mid,
   } = req.body;
-  // add validators to check that everything is a string
 
-  const queryString = `DELETE from watchlist WHERE userID = ${userID} and Movie_ID = ${mid}`;
+  // to check that all watchlist information is valid
 
-  connection.query(queryString, (error, results) => {
-    if (error) {
-      res.status(400).send('Database could not delete from watchlist');
-    } else {
-      res.send("Deleted selected watchlist entry");
-    }
-  });
+  const isValidUserID = userID !== undefined && typeof userID == "string";
+  const isValidMid = mid !== undefined && typeof mid == "string";
+
+  if (isValidUserID && isValidMid) {
+
+    const queryString = `DELETE from watchlist WHERE userID = ${userID} and Movie_ID = ${mid}`;
+
+    connection.query(queryString, (error, results) => {
+      if (error) {
+        res.status(400).send('Database could not delete from watchlist');
+      } else {
+        res.send("Deleted selected watchlist entry");
+      }
+    });
+  } else {
+    res.status(400).send("Invalid Input!");
+  };
 });
 
 module.exports = router;
